thema: First pass at [de]hydrate API
